+++ worker/include/RTC/RTCP/Packet.hpp
@@ -20,8 +20,8 @@ namespace RTC
 		// capture system but an estimation that moves. Refreshing it for audio less often
 		// than for video would make receivers combine measurements taken at different
 		// instants, and that difference is lip sync error.
-		constexpr uint16_t MaxAudioIntervalMs{ 1000 };
-		constexpr uint16_t MaxVideoIntervalMs{ 1000 };
+		constexpr int64_t MaxAudioIntervalMs{ 1000 };
+		constexpr int64_t MaxVideoIntervalMs{ 1000 };
 
 		enum class Type : uint8_t
 		{
